feat(security): isolate widget tick updates to non-exported receiver
This change mitigates a potential DoS vulnerability where the previous exported
`LinearClockWidgetReceiver` handled `ACTION_UPDATE_TICK`. An attacker could flood
this intent to trigger frequent widget updates.

Changes:
- Created `LinearClockTickReceiver` (non-exported) to handle tick updates.
- Updated `LinearClockWidgetReceiver` to delegate scheduling to the new receiver.
- Added migration logic in `LinearClockWidgetReceiver` to seamlessly transition existing alarms.
- Added legacy cleanup in `BootCompletedReceiver`.
- Updated `AndroidManifest.xml` to register the new receiver.
