+++ app/src/main/AndroidManifest.xml
@@ -42,6 +42,10 @@
         </activity>
 
         <receiver
+            android:name=".widget.LinearClockTickReceiver"
+            android:exported="false" />
+
+        <receiver
             android:name=".widget.LinearClockWidgetReceiver"
             android:exported="true"
             android:label="@string/widget_name">
