+++ app/src/main/java/com/dagsbalken/app/widget/LinearClockWidgetReceiver.kt
@@ -15,17 +15,26 @@ class LinearClockWidgetReceiver : GlanceAppWidgetReceiver() {
 
     override fun onEnabled(context: Context) {
         super.onEnabled(context)
-        scheduleUpdates(context)
+        LinearClockTickReceiver.scheduleUpdates(context)
     }
 
     override fun onDisabled(context: Context) {
         super.onDisabled(context)
-        cancelUpdates(context)
+        LinearClockTickReceiver.cancelUpdates(context)
+        // Attempt to cancel legacy updates if any exist
+        cancelLegacyUpdates(context)
     }
 
     override fun onReceive(context: Context, intent: Intent) {
         super.onReceive(context, intent)
+        // Migration: If we receive the legacy ACTION_UPDATE_TICK (from an old alarm that wasn't cancelled),
+        // we should migrate the user to the new non-exported receiver system immediately.
         if (intent.action == ACTION_UPDATE_TICK) {
+            // 1. Schedule the new secure updates
+            LinearClockTickReceiver.scheduleUpdates(context)
+            // 2. Cancel the old insecure updates so this receiver isn't triggered again
+            cancelLegacyUpdates(context)
+            // 3. Perform the update one last time to ensure continuity
             val pendingResult = goAsync()
             CoroutineScope(Dispatchers.Default).launch {
                 try {
@@ -40,10 +49,9 @@ class LinearClockWidgetReceiver : GlanceAppWidgetReceiver() {
     companion object {
         private const val ACTION_UPDATE_TICK = "com.dagsbalken.app.widget.ACTION_UPDATE_TICK"
         private const val REQUEST_CODE = 123
-        // Update every minute to keep the clock reasonably accurate
-        private const val UPDATE_INTERVAL_MILLIS = 60_000L
 
-        fun scheduleUpdates(context: Context) {
+        // Kept for legacy cleanup
+        fun cancelLegacyUpdates(context: Context) {
             val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as? AlarmManager ?: return
             val intent = Intent(context, LinearClockWidgetReceiver::class.java).apply {
                 action = ACTION_UPDATE_TICK
@@ -52,32 +60,12 @@ class LinearClockWidgetReceiver : GlanceAppWidgetReceiver() {
                 context,
                 REQUEST_CODE,
                 intent,
-                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
+                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE or PendingIntent.FLAG_NO_CREATE
             )
-
-            // Use setRepeating for efficiency. It's inexact on modern Android but ensures
-            // the widget updates roughly every minute without waking the device constantly.
-            // Using RTC so it relates to wall clock time.
-            alarmManager.setRepeating(
-                AlarmManager.RTC,
-                System.currentTimeMillis(),
-                UPDATE_INTERVAL_MILLIS,
-                pendingIntent
-            )
-        }
-
-        fun cancelUpdates(context: Context) {
-            val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as? AlarmManager ?: return
-            val intent = Intent(context, LinearClockWidgetReceiver::class.java).apply {
-                action = ACTION_UPDATE_TICK
+            if (pendingIntent != null) {
+                alarmManager.cancel(pendingIntent)
+                pendingIntent.cancel()
             }
-            val pendingIntent = PendingIntent.getBroadcast(
-                context,
-                REQUEST_CODE,
-                intent,
-                PendingIntent.FLAG_UPDATE_CURRENT or PendingIntent.FLAG_IMMUTABLE
-            )
-            alarmManager.cancel(pendingIntent)
         }
     }
 }
